localStorage update
every object that gets into localstorage now have an id and u will be able to delete single element from localstorage.

bug; in web console the application will get error from useEffect becuase its rendering multiple times because of not using " }, [];  array. but by not using this the page will not update itself.

What to fix this perhaps is passing the data to another compenent and then send it back.
Code code code..... :)